2-str_concat.c

+++ malloc_free/2-str_concat.c
@@ -22,8 +22,8 @@ char *str_concat(char *s1, char *s2)
 		s2 = "";
 	}
 
-	size_t len1 = strlen(s1);
-	size_t len2 = strlen(s2);
+	int len1 = strlen(s1);
+	int len2 = strlen(s2);
 
 	char *result = malloc(len1 + len2 + 1);
